requirements.txt
# PromptShiels Dependencies
# Core ML/DL
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.2.2
accelerate>=0.25.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Vector Database
faiss-cpu>=1.7.4

# Data Processing
pandas>=2.0.0
datasets>=2.14.0

# UI
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.1
